+++ thunder/tests/distributed/test_ddp.py
@@ -1,4 +1,3 @@
-import math
 import multiprocessing as mp
 import os
 import sys
@@ -50,105 +49,14 @@ if TE_AVAILABLE:
 
     is_fp8_supported, fp8_support_reason = check_fp8_support()
 
-try:
-    import expecttest  # noqa: F401
-    import hypothesis  # noqa: F401
-except ImportError:
-    raise ImportError(
-        "Required packages of `expecttest` and/or `hypothesis` are missing. "
-        "Install them with `pip install expecttest hypothesis`"
-    )
-from torch.testing._internal import common_distributed, common_utils
+from thunder.tests.distributed.helper import ToyModel, DataParallelTestCase
+from torch.testing._internal import common_utils
 
 executors_map = {
     TorchExecutor.name: TorchExecutor,
 }
 if nvFuserExecutor is not None:
     executors_map[nvFuserExecutor.name] = nvFuserExecutor
-
-
-# Compile - DDP tests
-def new_gelu(x):
-    return 0.5 * x * (1.0 + torch.tanh(math.sqrt(2.0 / math.pi) * (x + 0.044715 * torch.pow(x, 3.0))))
-
-
-class ToyModel(nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.net1 = nn.Linear(12, 12)
-        self.net2 = nn.Linear(12, 8)
-
-    def forward(self, x):
-        return self.net2(new_gelu(self.net1(x)))
-
-
-# note(crcrpar): How to write a test with `DDP`
-# Just add a method to :class:`CompileDDPTest`. The class is responsible for
-#     - calling `torch.distributed.init_process_group` with NCCL backend
-#     - setting rank to each process group / device
-# so what you'd need to do is to prepare a model and tensors, wrap the model with DDP, and
-# `thunder.jit` the original model or the DDP'd model, and do some computation and/or
-# examine the traces of the `thunder.jit`d.
-# If you force a test to be run with >2 GPUs for a test, you might want to inherit `CompileDDPTest`
-# and modify `world_size` to e.g. `max(torch.cuda.device_count(), 2)`.
-
-
-# note(crcrpar): Why inheriting `common_distributed.MultiProcessTestCase`?
-# When we're quite sure that we would only use `pytest` instead of `unittest`,
-# IIUC it's possible to run a test that is dependent on `DistributedDataParallel` and/or
-# `torch.distributed` by running the test file with [`torchrun`](https://pytorch.org/docs/stable/elastic/run.html),
-# but I don't think (a) it's quite intuitive to require `torchrun` explicitly to run a test and
-# (b) it's quite friendly to our CI as it's currently simply runs `pytest thunder/tests`.
-# I would say it's feasible to write a test with `torch.distributed` by using `torch.multiprocessing`,
-# but it would require us to make the function which defines the test logic picklable and would
-# lead to boilerplate test functions.
-# Ref: https://github.com/NVIDIA/apex/blob/7b2e71b0d4013f8e2f9f1c8dd21980ff1d76f1b6/apex/transformer/testing/distributed_test_base.py#L22
-class DataParallelTestCase(common_distributed.MultiProcessTestCase):
-    DISTRIBUTED_BACKEND = "nccl"
-
-    def __init__(self, *args, **kwargs) -> None:
-        super().__init__(*args, **kwargs)
-
-    def setUp(self) -> None:
-        super().setUp()
-        self._spawn_processes()
-
-    def tearDown(self) -> None:
-        torch.cuda.empty_cache()
-        super().tearDown()
-
-    # note(crcrpar): This means the world_size is up to two.
-    @property
-    def world_size(self) -> int:
-        return min(torch.cuda.device_count(), 2)
-
-    @property
-    def init_method(self):
-        return f"{common_utils.FILE_SCHEMA}{self.file_name}"
-
-    @classmethod
-    def _run(cls, rank, test_name, file_name, pipe):
-        self = cls(test_name)
-        self.rank = rank
-        self.file_name = file_name
-
-        torch.distributed.init_process_group(
-            init_method=self.init_method,
-            backend=self.DISTRIBUTED_BACKEND,
-            world_size=self.world_size,
-            rank=self.rank,
-        )
-
-        local_rank = self.rank % torch.cuda.device_count()
-        torch.cuda.set_device(local_rank)
-        os.environ["LOCAL_RANK"] = str(local_rank)
-
-        torch.distributed.barrier()
-        self.run_test(test_name, pipe)
-        torch.distributed.barrier()
-
-        torch.distributed.destroy_process_group()
-        sys.exit(0)
 
 
 @unittest.skipUnless(
@@ -1036,169 +944,6 @@ class CompileDDPTest(DataParallelTestCase):
         # notice how we cannot do `model.no_sync()` because it's not a ThunderModule
         with thunder.ThunderModule.no_sync(model):
             fwd_loss(model, x)
-
-    @pytest.mark.skipif(torch.cuda.device_count() < 2, reason="")
-    @common_utils.parametrize("name", ("column", "row"))
-    def test_tensor_parallel_linear(self, name):
-        device = torch.device("cuda", self.rank)
-        x = torch.randn(2, 12).to(device).requires_grad_()
-
-        process_group = None
-        ref_model = ToyModel().to(device)
-        with c10d._coalescing_manager(async_ops=True) as cm:
-            c10d.all_reduce(x)
-            for p in ref_model.parameters():
-                c10d.all_reduce(p)
-        cm.wait()
-
-        ref_state_dict = ref_model.state_dict()
-        expected = ref_model(x)
-
-        transform = {
-            "column": column_parallel,
-            "row": row_parallel,
-        }[name]
-        model = ToyModel().to(device)
-        model.load_state_dict(ref_state_dict)
-        jitted_model = thunder.jit(model)
-        tp_jitted_model = transform(
-            jitted_model,
-            target_modules=("net2",),
-            process_group=process_group,
-        )
-        y = tp_jitted_model(x)
-        torch.testing.assert_close(expected=expected, actual=y)
-
-        expected.mean().backward()
-        y.mean().backward()
-
-        dim = 1 if name == "row" else 0
-        expected_full_grad: torch.Tensor = ref_model.net2.weight.grad
-        expected = torch.chunk(expected_full_grad, self.world_size, dim)[self.rank]
-        torch.testing.assert_close(
-            expected=expected,
-            actual=tp_jitted_model.get_parameter("net2.weight").grad,
-        )
-
-    @pytest.mark.skipif(torch.cuda.device_count() < 2, reason="")
-    @common_utils.parametrize("name", ("column", "row"))
-    def test_tensor_parallel_embedding(self, name):
-        num_embeddings = 128
-        embedding_dim = 32
-
-        class Model(nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.embed = nn.Embedding(num_embeddings, embedding_dim)
-
-            def forward(self, x):
-                return self.embed(x)
-
-        device = torch.device(f"cuda:{self.rank}")
-        x = torch.randint(0, num_embeddings - 1, (16, 16), device=device)
-
-        process_group = None
-        ref_model = Model().to(device)
-        with c10d._coalescing_manager(async_ops=True) as cm:
-            for p in ref_model.parameters():
-                c10d.all_reduce(p)
-        cm.wait()
-
-        ref_state_dict = ref_model.state_dict()
-        expected = ref_model(x)
-
-        transform = {
-            "column": column_parallel,
-            "row": row_parallel,
-        }[name]
-        model = Model().to(device)
-        model.load_state_dict(ref_state_dict)
-        jitted_model = thunder.jit(model)
-        tp_jitted_model = transform(
-            jitted_model,
-            target_modules=("embed",),
-            process_group=process_group,
-        )
-        y = tp_jitted_model(x)
-
-        dim: int
-        orig_size: int
-        if name == "column":
-            dim = 0
-            orig_size = num_embeddings
-        else:
-            dim = 1
-            orig_size = embedding_dim
-        torch.testing.assert_close(
-            tp_jitted_model.get_parameter("embed.weight").size(dim), orig_size // self.world_size
-        )
-        torch.testing.assert_close(expected=expected, actual=y)
-
-        expected.mean().backward()
-        y.mean().backward()
-
-        torch.testing.assert_close(
-            expected=ref_model.embed.weight.grad.chunk(self.world_size, dim)[self.rank],
-            actual=tp_jitted_model.get_parameter("embed.weight").grad,
-        )
-
-    # TODO(crcrpar): Activate numerical check
-    @pytest.mark.skipif(torch.cuda.device_count() < 2, reason="")
-    def test_tensor_parallel_both_column_and_row(self):
-        num_embeddings = 128
-        embedding_dim = 32
-        n_hidden = 96
-        n_out = 16
-
-        class Model(nn.Module):
-            def __init__(self):
-                super().__init__()
-                self.embed_1 = nn.Embedding(num_embeddings, embedding_dim)
-                self.embed_2 = nn.Embedding(num_embeddings, embedding_dim)
-                self.linear1_0 = nn.Linear(embedding_dim, n_hidden)
-                self.linear1_1 = nn.Linear(n_hidden, n_hidden)
-                self.linear2_0 = nn.Linear(n_hidden, n_hidden)
-                self.linear2_1 = nn.Linear(n_hidden, n_out)
-
-            def forward(self, x):
-                feat_1 = self.embed_1(x)
-                feat_2 = self.embed_2(x)
-                sum_of_feat = feat_1 + feat_2
-                h = self.linear1_1(self.linear1_0(sum_of_feat))
-                return self.linear2_1(self.linear2_0(h))
-
-        device = torch.device("cuda", self.rank)
-        x = torch.randint(0, num_embeddings - 1, (16, 16), device=device)
-
-        process_group = None
-        ref_model = Model().to(device)
-        ref_state_dict = ref_model.state_dict()
-        expected = ref_model(x)
-
-        model = Model().to(device)
-        model.load_state_dict(ref_state_dict)
-        jitted_model = thunder.jit(model)
-
-        column_parallels = ["embed_1", "linear1_0", "linear2_1"]
-        row_parallels = ["embed_2", "linear1_1", "linear2_0"]
-        tp_jitted_model = row_parallel(
-            column_parallel(
-                jitted_model,
-                column_parallels,
-                process_group,
-            ),
-            row_parallels,
-            process_group,
-        )
-        actual = tp_jitted_model(x)
-        actual.mean().backward()
-
-        for col, orig_size in zip(column_parallels, [num_embeddings, n_hidden, n_out]):
-            weight = tp_jitted_model.get_parameter(f"{col}.weight")
-            self.assertEqual(weight.size(0), orig_size // self.world_size)
-        for row, orig_size in zip(row_parallels, [embedding_dim, n_hidden, n_hidden]):
-            weight = tp_jitted_model.get_parameter(f"{row}.weight")
-            self.assertEqual(weight.size(1), orig_size // self.world_size)
 
 
 common_utils.instantiate_parametrized_tests(CompileDDPTest)
